+++ src/events/events.ts
@@ -1,0 +1,30 @@
+import { ComputeSkipReason } from "ton-core";
+import { TrackedMessage } from "./message";
+
+export type TrackedEvent = {
+    type: 'deploy'
+} | {
+    type: 'received',
+    message: TrackedMessage
+} | {
+    type: 'received-bounced',
+    message: TrackedMessage
+} | {
+    type: 'failed',
+    errorCode: number,
+    errorMessage?: string
+} | {
+    type: 'processed',
+    gasUsed: bigint
+} | {
+    type: 'skipped',
+    reason: ComputeSkipReason
+} | {
+    type: 'sent',
+    messages: TrackedMessage[]
+} | {
+    type: 'sent-bounced',
+    message: TrackedMessage
+} | {
+    type: 'sent-bounced-failed'
+};